Update Project tracker columns to emoji
On the "First few days" page the advice to emoji the "Awaiting Start"
column in the Project tracker didn't match that on the Checklist or
Project Tracking pages, where the advice is to emoji the "Awaiting
go/no-go" column. This change aligns the three to all refer to the
"Awaiting go/no-go" column.

Advice to look at the Service areas list has also been added.